refactor(ds, apps): update components
